+++ g419-crete-api/src/main/java/g419/crete/api/trainer/WekaJ48ClusterMentionTrainer.java
@@ -4,10 +4,10 @@ import g419.crete.api.instance.ClusterClassificationInstance;
 import weka.classifiers.Classifier;
 import weka.core.Instance;
 
-public class WekaJ48ClusterMentionTrainer extends AbstractCreteTrainer<Classifier, ClusterClassificationInstance, Instance, Integer>{
+public class WekaJ48ClusterMentionTrainer extends AbstractCreteTrainer<Classifier, ClusterClassificationInstance, Instance, Double>{
 
 	@Override public Class<Classifier> getModelClass() {return Classifier.class;}
 	@Override public Class<ClusterClassificationInstance> getAbstractInstanceClass() {return ClusterClassificationInstance.class;}
 	@Override public Class<Instance> getClassifierInstanceClass() {return Instance.class;}
-	@Override public Class<Integer> getLabelClass() {	return Integer.class;}
+	@Override public Class<Double> getLabelClass() {	return Double.class;}
 }
